--- a/crates/shared/ab-proof-of-space/src/lib.rs
+++ b/crates/shared/ab-proof-of-space/src/lib.rs
@@ -16,6 +16,7 @@
     sync_unsafe_cell,
     vec_into_raw_parts
 )]
+extern crate alloc;
 
 pub mod chia;
 pub mod chiapos;