This is a follow-up to https://github.com/nazar-pc/abundance/pull/380, where all allocations are fixed size. The performance is mostly the same as before with some variability. While I did notice slightly higher peak performance, I'd say it is inconclusive:
```
Before:
chia/table/single/1x    time:   [750.41 ms 760.79 ms 772.80 ms]
                        thrpt:  [1.2940  elem/s 1.3144  elem/s 1.3326  elem/s]
chia/table/parallel/8x  time:   [543.21 ms 549.05 ms 553.93 ms]
                        thrpt:  [14.442  elem/s 14.571  elem/s 14.727  elem/s]
After:
chia/table/single/1x    time:   [747.82 ms 756.94 ms 768.09 ms]
                        thrpt:  [1.3019  elem/s 1.3211  elem/s 1.3372  elem/s]
chia/table/parallel/8x  time:   [529.15 ms 534.42 ms 540.15 ms]
                        thrpt:  [14.811  elem/s 14.969  elem/s 15.119  elem/s]
```

It is nice to break 15 elements/s on 8C16T CPU though.

The biggest noticeable change is probably even lower memory usage with `y`s of the parent table being freed after use like metadata clearing introduced in https://github.com/nazar-pc/abundance/pull/380.